+++ src/dump.c
@@ -431,7 +431,7 @@ dump_screenshot(void)
 {
     char screenshot[BUFSZ];
     char *filename = get_dump_filename();
-    Sprintf(screenshot, "%s_screenshot_%ld_t%ld.html", filename, u.ubirthday, moves);
+    Sprintf(screenshot, "%s_screenshot_%ld_t%ld.html", filename, (long)u.ubirthday, moves);
     if (filename) free(filename);
 
     html_dump_fp = fopen(screenshot, "w");
